Fix refraction (works for n = 1)

+++ src/vector_ops.c
@@ -6,7 +6,7 @@
 /*   By: ede-thom <ede-thom@42.edu.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/11/24 18:36:49 by dhorvill          #+#    #+#             */
-/*   Updated: 2020/06/09 18:07:55 by ede-thom         ###   ########.fr       */
+/*   Updated: 2020/06/16 14:09:51 by ede-thom         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -113,7 +113,7 @@ float	distance(t_point p1, t_point p2)
 	return (sqrt(pow(p2.x - p1.x, 2) + pow(p2.y - p1.y, 2) + pow(p2.z - p1.z, 2)));
 }
 
-float	angle(t_vect v1, t_vect v2)
+double	angle(t_vect v1, t_vect v2)
 {
 	return(acos(dot(v1, v2) / (norm(v1) * norm(v2))));
 }
